modify edgeshift datatype

+++ src/pyggnn/data/dataset.py
@@ -58,7 +58,7 @@ class BaseGraphDataset(torch.utils.data.Dataset):
             edge_index=torch.stack(
                 [torch.LongTensor(edge_src), torch.LongTensor(edge_dst)], dim=0
             ),
-            edge_shift=torch.tensor(edge_shift),
+            edge_shift=torch.tensor(edge_shift, dtype=torch.float64),
         )
         return data
